Bug 1787500 - Split the main code into modules and web components for clearer code separation.
