+++ extension/content/byEngineView.css
@@ -1,0 +1,33 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+#engine-selection {
+  display: grid;
+  grid: auto-flow / max-content 1fr 1fr 1fr;
+  align-items: center;
+}
+
+#engine-selection > div {
+  padding: 0 0.5em;
+}
+
+#locale-region-results {
+  display: grid;
+  grid: auto-flow / max-content 1fr;
+}
+
+#locale-region-results > div {
+  padding: .3em 1em;
+}
+
+#locale-region-results > div:nth-child(-n+2) {
+  font-weight: bold;
+  text-align: center;
+  padding-bottom: .5em;
+}
+
+#locale-region-results > div:nth-child(2n+3):nth-child(4n+3),
+#locale-region-results > div:nth-child(2n+4):nth-child(4n+4) {
+  background: var(--in-content-box-background-odd);
+}
