+++ extension/content/style.css
@@ -1,3 +1,7 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 @import url("chrome://global/skin/in-content/common.css");
 
 html {
@@ -115,68 +119,6 @@ h1, h2, h3, h4 {
   display: block;
 }
 
-#locale-region-settings {
-  display: grid;
-  justify-content: space-between;
-  align-items: center;
-  width: max-content;
-  margin: 0 auto;
-  grid-template-columns: repeat(3, auto);
-  grid-template-rows: repeat(2, auto);
-}
-
-#region-select-wrapper,
-#locale-select-wrapper {
-  grid-row: 1 / 3;
-}
-
-#locale-region-settings div {
-  margin-right: 2em;
-  text-align: center;
-}
-
-#locale-region-settings label {
-  display: block;
-}
-
-#engines-table {
-  display: grid;
-  grid-template-columns: repeat(5, max-content) auto;
-  padding-top: 1em;
-}
-
-#engines-table > div {
-  padding: .3em 1em;
-}
-
-#engines-table > div:nth-child(-n+6) {
-  font-weight: bold;
-  text-align: center;
-  padding-bottom: .5em;
-}
-
-#engines-table > div:nth-child(n+6) {
-  opacity: 0.8;
-  cursor: pointer;
-}
-
-#engines-table > div:nth-child(n+6):hover {
-  opacity: 1;
-}
-
-#engines-table > div:nth-child(6n+7):nth-child(12n+7),
-#engines-table > div:nth-child(6n+8):nth-child(12n+8),
-#engines-table > div:nth-child(6n+9):nth-child(12n+9),
-#engines-table > div:nth-child(6n+10):nth-child(12n+10),
-#engines-table > div:nth-child(6n+11):nth-child(12n+11),
-#engines-table > div:nth-child(6n+12):nth-child(12n+12) {
-  background: var(--in-content-box-background-odd);
-}
-
-#private-browsing-engine {
-  color: var(--in-content-text-color);
-}
-
 .params {
   white-space: nowrap;
   overflow: hidden;
@@ -235,54 +177,6 @@ fieldset > div td {
 #reload-div > button {
   margin-right: auto;
   margin-left: auto;
-}
-
-#engine-selection {
-  display: grid;
-  grid: auto-flow / max-content 1fr 1fr 1fr;
-  align-items: center;
-}
-
-#engine-selection > div {
-  padding: 0 0.5em;
-}
-
-#locale-region-results {
-  display: grid;
-  grid: auto-flow / max-content 1fr;
-}
-
-#locale-region-results > div {
-  padding: .3em 1em;
-}
-
-#locale-region-results > div:nth-child(-n+2) {
-  font-weight: bold;
-  text-align: center;
-  padding-bottom: .5em;
-}
-
-#locale-region-results > div:nth-child(2n+3):nth-child(4n+3),
-#locale-region-results > div:nth-child(2n+4):nth-child(4n+4) {
-  background: var(--in-content-box-background-odd);
-}
-
-#overview {
-  display: flex;
-}
-
-#overview > label {
-  margin: auto 0;
-  padding: 1ch;
-}
-
-#diff-display {
-  margin: 1ch 0;
-  font-size: 12px;
-}
-
-#diff-display > pre {
-  margin: 0;
 }
 
 #config-error {
