+++ extension/content/utils.mjs
